+++ NEWS.md
@@ -1,3 +1,8 @@
+# rrapply 1.2.6
+
+* Duplicate column names are allowed in `how = "bind"` and made unique with `make.unique()`;
+* Fixed minor issue in `how = "names"` when `f` returns length-zero characters
+
 # rrapply 1.2.5
 
 * **Breaking change**: data.frame columns in `how = "bind"` only include child list names